Tidy up hisim_main.py and console output
