+++ hisim/hisim_main.py
@@ -10,12 +10,12 @@ from typing import Optional, Any, cast
 import argparse
 from pydantic import TypeAdapter
 from dotenv import load_dotenv
+
 try:
+    from hisim.json_executor import setup_components_and_connections
     from hisim.postprocessingoptions import PostProcessingOptions
-    from hisim.components.generic_car import GenericCarInformation
     import hisim.simulator as sim
-    from hisim import log, utils
-    from hisim import loadtypes as lt
+    from hisim import log
     from hisim.simulationparameters import SimulationParameters
 except ModuleNotFoundError:
     raise ModuleNotFoundError(
@@ -23,13 +23,6 @@ except ModuleNotFoundError:
         "It may not be installed in the current Python environment.\n\n"
         "If you already installed HiSim locally with 'pip install -e .', "
         "make sure you are using the same virtual environment/interpreter."
-    ) from None
-try:
-    import humps
-except ModuleNotFoundError:
-    raise ModuleNotFoundError(
-        "The 'humps' library is required for JSON-based scenario execution. "
-        "Please install it with 'pip install pyhumps==3.8.0'."
     ) from None
 
 load_dotenv()
@@ -143,165 +136,7 @@ def initialize_from_json(
         my_simulation_parameters=sim_params,
     )  # type: ignore[no-any-return]
 
-    # Build Components
-    components = scenario_data.get("components", [])
-    if not components:
-        raise ValueError("No components defined in scenario")
-
-    component_dict = {}
-
-    def import_from_string(full_classname: str):
-        """Import a class from a full module path string."""
-        # This function was created with the help of ChatGPT
-
-        try:
-            module_path, class_name = full_classname.rsplit(".", 1)
-        except ValueError as e:
-            raise ValueError(f"Invalid class path: {full_classname}") from e
-
-        try:
-            module = importlib.import_module(module_path)
-        except ModuleNotFoundError as e:
-            raise ImportError(f"Could not import module '{module_path}'") from e
-
-        try:
-            return getattr(module, class_name)
-        except AttributeError as e:
-            raise ImportError(f"Module '{module_path}' has no class '{class_name}'") from e
-
-    for comp_def in components:
-        component_class = import_from_string(comp_def["component_full_classname"])
-        config_class = import_from_string(comp_def["config_full_classname"])
-        try:
-            config_dict = comp_def.get("configuration", {})
-            # Fill in the absolute paths we have filled with placeholders in JSON
-            if comp_def["component_full_classname"] == "hisim.components.loadprofilegenerator_utsp_connector.UtspLpgConnector":
-                if config_dict["result_dir_path"] == "<<utils.HISIMPATH['utsp_results']>>":
-                    config_dict["result_dir_path"] = utils.HISIMPATH["utsp_results"]
-                else:
-                    # This warning was generated by Copilot
-                    log.warning(f"Unexpected value for result_dir_path in UtspLpgConnector config: {config_dict['result_dir_path']}.")
-                # UtspLpgConnectorConfig has values of type JsonReference, which need to be converted to camel-case
-                config_dict["household"] = humps.pascalize(config_dict["household"])
-                config_dict["travel_route_set"] = humps.pascalize(config_dict["travel_route_set"])
-                config_dict["transportation_device_set"] = humps.pascalize(config_dict["transportation_device_set"])
-                config_dict["charging_station_set"] = humps.pascalize(config_dict["charging_station_set"])
-            elif comp_def["component_full_classname"] == "hisim.components.weather.Weather":
-                if "<<utils.get_input_directory()>>" in config_dict["source_path"]:
-                    config_dict["source_path"] = config_dict["source_path"].replace("<<utils.get_input_directory()>>", utils.get_input_directory())
-                else:
-                    # This warning was generated by Copilot
-                    log.warning(f"Unexpected value for source_path in Weather config: {config_dict['source_path']}.")
-
-            # Use the JSONWizard from_dict method to get ConfigBase instance
-            config = config_class.from_dict(config_dict)
-
-            if comp_def["component_full_classname"] == "hisim.components.generic_car.Car":
-                # We have to generate the car_info_dict
-                car_info = None
-                found = False
-                for comp in my_sim.wrapped_components:
-                    if comp.my_component.get_full_classname() == "hisim.components.loadprofilegenerator_utsp_connector.UtspLpgConnector":
-                        if config_dict["name"] in comp.my_component.config.cars:
-                            found = True
-                            car_info = GenericCarInformation(comp.my_component).data_dict_for_car_component[config_dict["household_name"]]
-                if not found:
-                    raise ValueError(f"The car '{config_dict['name']}' was not associated with any UTSP connector.")
-                component = component_class(
-                    config=config,
-                    my_simulation_parameters=sim_params,
-                    data_dict_with_car_information=car_info,
-                )
-            else:
-                component = component_class(
-                    config=config,
-                    my_simulation_parameters=sim_params,
-                )
-        except TypeError as e:
-            raise ValueError(
-                f"Failed to initialize component {component_class.__name__}: {e}"
-            ) from e
-
-        # Create inputs/outputs
-        for input_def in comp_def.get("inputs", []):
-            if input_def["dynamic"]:
-                component.add_component_input_and_connect(
-                    source_component_output=input_def["source_component_output"],
-                    source_object_name=input_def["source_object_name"],
-                    source_load_type=lt.LoadTypes(input_def["source_load_type"]),
-                    source_unit=lt.Units(input_def["source_unit"]),
-                    source_tags=input_def["source_tags"],
-                    source_weight=input_def["source_weight"],
-                )
-            else:
-                component.add_input(
-                    object_name=input_def["object_name"],
-                    field_name=input_def["field_name"],
-                    load_type=lt.LoadTypes(input_def["load_type"]),
-                    unit=lt.Units(input_def["unit"]),
-                    mandatory=input_def["mandatory"],
-                )
-        for output_def in comp_def.get("outputs", []):
-            if output_def["dynamic"]:
-                component.add_component_output(
-                    source_output_name=output_def["source_output_name"],
-                    source_tags=output_def["source_tags"],
-                    source_load_type=lt.LoadTypes(output_def["source_load_type"]),
-                    source_unit=lt.Units(output_def["source_unit"]),
-                    source_weight=output_def["source_weight"],
-                    output_description=output_def["output_description"],
-                    source_component_class=output_def.get("source_component_class", None),
-                )
-            else:
-                component.add_output(
-                    object_name=output_def["object_name"],
-                    field_name=output_def["field_name"],
-                    load_type=lt.LoadTypes(output_def["load_type"]),
-                    unit=lt.Units(output_def["unit"]),
-                    postprocessing_flag=output_def.get("postprocessing_flag", None),
-                    sankey_flow_direction=output_def.get("sankey_flow_direction", None),
-                    output_description=output_def.get("output_description", None),
-                )
-
-        my_sim.add_component(component, connect_automatically=comp_def["connect_automatically"])
-        component_dict[component.config.name] = component
-
-    # Connect components
-    for conn in scenario_data.get("connections", []):
-        try:
-            source_name = conn["source"]["component_name"]
-            source_field = conn["source"]["field_name"]
-
-            target_name = conn["target"]["component_name"]
-            target_field = conn["target"]["field_name"]
-
-            # Resolve components
-            source_comp = component_dict.get(source_name)
-            target_comp = component_dict.get(target_name)
-            if source_comp is None:
-                log.warning(f"Source component not found: {source_name}")
-                continue
-            if target_comp is None:
-                log.warning(f"Target component not found: {target_name}")
-                continue
-
-            # Connect (analogous to Python-based setups)
-            try:
-                target_comp.connect_input(
-                    target_field,
-                    source_name,
-                    source_field,
-                )
-                log.information(
-                    f"Added connection: {source_name}.{source_field} -> {target_name}.{target_field}"
-                )
-
-            except Exception as e:
-                # Likely, fields named like Input_..._value_number do not actually have a corresponding input?!
-                log.warning(f"Failed to connect {source_name}.{source_field} to {target_name}.{target_field}: {e}")
-
-        except KeyError as e:
-            raise ValueError(f"Malformed connection entry: missing {e}") from e
+    my_sim = setup_components_and_connections(scenario_data, my_sim, sim_params)
 
     if delta:
         # delta_data = load_json_file(delta)
@@ -315,9 +150,9 @@ def initialize_from_json(
 def run_simulation(my_sim: sim.Simulator, path_to_module: Optional[str]) -> None:
     """Runs the simulation (for both Python-based and JSON-based executions)."""
 
-    # Print components and their inputs/outputs
+    # If debugging is needed, this may be used to print components and their inputs/outputs
     for comp in my_sim.wrapped_components:
-        log.information(f"Component {comp.my_component.component_name} has inputs "
+        log.debug(f"Component {comp.my_component.component_name} has inputs "
                         f"{[input.full_name for input in comp.component_inputs]} and"
                         f" outputs {[output.full_name for output in comp.component_outputs]}")
 
@@ -456,6 +291,8 @@ def main_cli():
                 except Exception:
                     log.information("Logging default file could not be removed. This can occur when more than one simulation run simultaneously.")
 
+        my_sim: sim.Simulator
+        ptm: str
         # Dispatching logic
         if config["mode"] == "python":
             print(f"Calling setup_function from {config['module_file']}")
